seperate schema of table and sample data seeding into two files
